yakunote-backend/requirements.txt
fastapi
uvicorn
cohere
openai
python-dotenv
supabase
trafilatura
requests
